database/pyproject.toml
[project]
name = "llm-interactions-database"
version = "0.1.0"
description = "Database scripts and utilities for LLM Interactions"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
explicit_package_bases = true
mypy_path = ".."
namespace_packages = true
exclude = ["tests"]

[[tool.mypy.overrides]]
module = "psycopg2.*"
ignore_missing_imports = true

[dependency-groups]
dev = [
    "types-psycopg2>=2.9.21.20251012",
]
